third_party/WebKit/LayoutTests/external/wpt/payment-request/allowpaymentrequest/no-attribute-same-origin-bc-containers.https-expected.txt

This is a testharness.js-based test.
FAIL iframeassert_equals: expected "Success" but got "Exception"
FAIL frameassert_equals: expected "Success" but got "Exception"
FAIL objectassert_equals: expected "Success" but got "Exception"
FAIL embedassert_equals: expected "Success" but got "Exception"
Harness: the test ran to completion.
